PLUGINAPI-44 Moved testers out of the library, to the dedicated test-fixtures artifact
