+++ test-fixtures/src/main/java/org/sonar/api/testfixtures/measure/TestMeasure.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.api.ce.measure.test;
+package org.sonar.api.testfixtures.measure;
 
 import javax.annotation.concurrent.Immutable;
 import org.sonar.api.ce.measure.Measure;
