+++ test-fixtures/src/test/java/org/sonar/api/testfixtures/measure/TestMeasureTest.java
@@ -17,9 +17,10 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.api.ce.measure.test;
+package org.sonar.api.testfixtures.measure;
 
 import org.junit.Test;
+import org.sonar.api.testfixtures.measure.TestMeasure;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
